+++ transit_to_gif_handlers.py
